% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATmv.R
\name{MATmv}
\alias{MATmv}
\title{MATmv: Manejo de datos con casos completos y NA}
\usage{
MATmv(dataframe, ...)
}
\arguments{
\item{dataframe}{Un dataframe que contiene las variables.}

\item{...}{Variables a seleccionar. Pueden ingresarse directamente sin comillas (ej. mpg, hp, wt).}
}
\value{
El nombre (character) del nuevo dataframe con los casos completos: "\if{html}{\out{<dataframe>}}_sm".
}
\description{
Esta función selecciona variables de un dataframe, identifica casos completos y crea un resumen
con información sobre los datos faltantes. Si no se especifican variables, se utilizan todas las columnas.
}
\details{
La función genera:
\enumerate{
\item Un nuevo dataframe con casos completos (según las columnas seleccionadas) en el entorno global:
\itemize{
\item "\if{html}{\out{<dataframe>}}_sm"
}
\item Una lista de información también en el entorno global:
\itemize{
\item "\if{html}{\out{<dataframe>}}_sm_info"
Contiene:
\itemize{
\item tabla_na: tabla HTML con los casos con NA (o NULL si no hay)
\item grafico_vis_miss: gráfico de visdat::vis_miss para las columnas seleccionadas
}
}
}
}
\examples{
# Usar todas las variables del dataframe:
MATmv(mtcars)

# Usar varias variables sin comillas:
MATmv(mtcars, mpg, hp, wt)

}
