% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATtaf.R
\name{MATtaf}
\alias{MATtaf}
\title{MATtaf: Tabla de frecuencias agrupadas en intervalos}
\usage{
MATtaf(data, variable, breaks = NULL)
}
\arguments{
\item{data}{Dataframe que contiene los datos.}

\item{variable}{Variable numérica para la que se calculará la distribución de frecuencias (sin comillas).}

\item{breaks}{(Opcional) Número de intervalos deseados. Si no se especifica, se calcula automáticamente con Freedman-Diaconis.}
}
\value{
Una lista con la tabla de frecuencias y un histograma, guardada en el entorno global
con el nombre "\if{html}{\out{<variable>}}_intervalos_frecuencia". Invisiblemente devuelve ese nombre.
}
\description{
Esta función crea una tabla con la distribución de frecuencias agrupadas en intervalos
para una variable numérica de un dataframe. Si no se especifica el número de intervalos,
se utiliza la regla de Freedman-Diaconis para determinarlo automáticamente. Además, genera un histograma
cuyos intervalos coinciden con los de la tabla.
}
\examples{
\dontrun{
MATtaf(data = mtcars, variable = mpg)
# Resultados: mpg_intervalos_frecuencia en el entorno global
}

}
